Bulk-fix K&R default-argument-promotion type conflicts
A K&R-declared (unspecified-args) forward declaration in functions.h
can't be defined with a default-promoted parameter type (byte/char/
short/float) - such args are promoted to int (or double, for float)
when passed anyway, so this changes nothing about the actual calling
convention. Same fix already applied by hand to Shutdown.c and
CGameState_ScalarDeletingDestructor.c; this batch applies it
programmatically across every raw-ported function whose own
definition-line parameter list used one of these types, fixing 169
files (138 now confirmed compiling clean via winegcc-stable -c).

+++ src/network/HandleTurnTimeoutSlot.c
@@ -12,7 +12,7 @@
 #include <windows.h>
 
 
-void HandleTurnTimeoutSlot(byte param_1)
+void HandleTurnTimeoutSlot(int param_1)
 
 {
   byte *pbVar1;
