+++ src/unnamed/FUN_0041ea30.c
@@ -8,7 +8,7 @@
 #include "ghidra_types.h"
 
 
-uint FUN_0041ea30(undefined4 param_1,ushort param_2)
+uint FUN_0041ea30(undefined4 param_1,int param_2)
 
 {
   uint uVar1;
