+++ src/unnamed/FUN_0048f260.c
@@ -7,7 +7,7 @@
 #include "ghidra_types.h"
 
 
-void * __thiscall FUN_0048f260(void *param_1,byte param_2)
+void * __thiscall FUN_0048f260(void *param_1,int param_2)
 
 {
   FUN_0048f280(param_1);
