+++ src/unnamed/FUN_004ec430.c
@@ -10,7 +10,7 @@
 
 /* WARNING: Globals starting with '_' overlap smaller symbols at the same address */
 
-void __thiscall FUN_004ec430(int param_1,int param_2,int param_3,byte param_4,uint param_5)
+void __thiscall FUN_004ec430(int param_1,int param_2,int param_3,int param_4,uint param_5)
 
 {
   float fVar1;
